+++ core/metadata/src/impls/metadata.rs
@@ -293,7 +293,7 @@ where
 
         // TODO: calculate the index;
         let idx = header.op as usize;
-        assert_eq!(header.command, Command2::Prepare);
+        assert_eq!(header.command, Command2::Prepare as u8);
         assert!(
             journal.handle().header(idx).is_none(),
             "replicate: must not already have prepare"
